Use fetch content for google tests. Remove interface due to overhead on embedded platforms. Simplify file structure.
