--- .gitmodules
@@ -1,3 +1,0 @@
-[submodule "external/googletest"]
-	path = external/googletest
-	url = https://github.com/google/googletest.git
