--- external/CMakeLists.txt
@@ -1,4 +1,0 @@
-set(CMAKE_CXX_STANDARD 20)
-
-# Add GTests to project
-add_subdirectory(googletest)
