+++ tests/CMakeLists.txt
@@ -1,3 +1,16 @@
+include(FetchContent)
+
+# Fetch googletests
+FetchContent_Declare(
+  googletest
+  GIT_REPOSITORY https://github.com/google/googletest.git
+  GIT_TAG        v1.14.0  # Use any tag, commit hash, or branch
+)
+
+set(gtest_force_shared_crt ON CACHE BOOL "" FORCE) # For windows
+
+FetchContent_MakeAvailable(googletest)
+
 add_executable(simplePIDControllerTests 
     simplePIDControllerTests.cpp 
 )
